requirements.txt
streamlit-mic-recorder
InstructorEmbedding
sentence_transformers
ctransformers
transformers
langchain
langchain_community
streamlit
pypdfium2
librosa
pyyaml
torch
llama-cpp-python
pydantic
Chromadb
bs4